// src/app/swaggerFiles/model/role.ts
/**
 * assoparenteleve API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type Role = 'SUPER_ADMIN' | 'ADMIN' | 'PARENT' | 'PROFESSOR';

export const Role = {
    SUPERADMIN: 'SUPER_ADMIN' as Role,
    ADMIN: 'ADMIN' as Role,
    PARENT: 'PARENT' as Role,
    PROFESSOR: 'PROFESSOR' as Role
};